requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
streamlit==1.28.2
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.2 